Address review

+++ core/src/upload.rs
@@ -90,7 +90,7 @@ pub async fn upload(
 #[derive(Serialize, Deserialize, Tsify)]
 #[tsify(into_wasm_abi, from_wasm_abi)]
 pub struct PreviewUploadRequest {
-    pub size: i32,
+    pub size: i64,
 }
 
 #[derive(Serialize, Deserialize, Tsify)]
